+++ Tree Tracker/Services/Api.swift
@@ -1,9 +1,0 @@
-import Foundation
-import Alamofire
-import class UIKit.UIImage
-
-protocol Api {
-    @available(*, deprecated, message: "Replaced by ProtectEarthTreeService")
-    func upload(tree: LocalTree, progress: @escaping (Double) -> Void, completion: @escaping (Result<AirtableTree, AFError>) -> Void) -> Cancellable
-    func loadImage(url: String, completion: @escaping (UIImage?) -> Void)
-}